requirements.txt
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
idna==2.8
mysql==0.0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
PyMySQL==0.9.3
pyTelegramBotAPI==3.6.7
requests==2.22.0
six==1.13.0
telebot==0.0.3
urllib3==1.25.6
psycopg2-binary==2.8.4


